Added the InputValidator class
index.jsp and newSite.jsp now validate the input passed to them and displays any errors the user made.
newSite shouldn't return any database error now, as the inputs are sanitised.

It should be noted that the site is still vulnerable to some types of SQL injection attack, but it should block the most dangerous ones.